Disabled usage data collection
This usage data (user name, app name, etc.) needs to be collected when
these apps are run inside Sony's network. The logging to our server
fails gracefully when the client's computer is not connected to Sony's
network. With this change, we no longer collect this data in local
memory. Added a 'PUBLIC' pre-processing symbol.