fixed bug with refreshing properties
PropertyGridView was refreshing the currently selected
PropertyEditingControl in response to a paint message. This could cause
the user to lose their currently edited value, like for the Name
property of circuit elements in the Circuit Editor. This change reverts
the logic to be like it was before ATF 3.8 was released.