Perforce P4API.NET update and clean-up
Tried to address bug #31 about there being many Perforce server
processes left running. I was never able to reproduce this problem with
the Windows version of the Perforce server.

1. Updated the P4API.NET library from Perforce from 2014.1.83.7625 to
2014.2.96.821, to get this bug fix in particular:
http://www.perforce.com/perforce/doc.current/user/p4api.netnotes.txt
A memory leak has been fixed which could previously cause
crashes in client applications built with the .NET API.
Connections are now properly closed on the server side after
a command has run.
2. Made sure that all Perforce-related objects are disposed of
correctly. Previously, the command objects were never disposed and the
connection object was not disposed when the app shut down.
3. Added the ApplicationName property to PerforceService so that
connections to Perforce can be named appropriately. Be default, it will
be the application's file name.
4. Did some minor clean-up and refactoring.