WPF windows docking framework fixes
Fixed an edge case with WPF layout to prevent the app from being
restored to a position offscreen if the user reconfigured the layout of
their monitors before relaunching.
Fixed a WPF windows layout restoring issue if an app was running on the
secondary monitor.
Fixed an issue in the WPF docking framework where if the user tried to
drag a maximized floating window, it could be positioned offscreen when
restored to a non-maximized state.
Fixed an issue in the WPF docking framework where floating windows did
not handle shortcut keys.