SkinService will now re-use property objects
The SkinService applies the skin by setting properties on C# objects.
Previously, if the property was a reference type, a new object was
always created. This could cause problems if the object already existed
and was a derived class or had been configured in some special way
outside the skin file. Now, if the property's object exists already
(i.e., the getter does not return null), it will be reused unless the
skin file specifies a constructor with parameters. Replaced an exception
with a warning message, so that skin files can specify properties on
objects that don't exist.