ControlHostService could fail to show context menu
ControlHostService: fixed a bug where the context menu on a panel's tab
might not be displayed when the user right-clicks on the tab, until the
user had changed focus to another tab and then back again. This was
found by a client because their app loaded a document and then
immediately set the Windows docking framework layout.