FormNcRenderer: fixed a crash that could happen if an app used multiple GUI threads
This class is for custom skinning of non-client areas of a Window. There
was an assumption that the .NET Application.ApplicationExit event would
only be fired once, but with multiple GUI threads, if one of those
threads exits while others are running, this event could be fired
multiple times and was causing a resource to be disposed of prematurely.
This resource didn't actually need to be disposed of.