tabbed Control focus event not always happening
Fixed a bug where the client's IControlHostClient.Activate method was
not always called in response to a mouse-down event on a tabbed Control.
DockPanelSuite seemed to be filtering out these messages when the
DockPane (a collection of client-registered Controls in a tab group) was
activated already, even though the user clicked on a different tab.