Implemented Dynamic Properties in CircuitEditor
CircuitEditor sample app: added a dynamic property system. This lets the
user add, edit, and remove properties to any circuit element, using new
context menu commands on the property editor.

CircuitEditingContext: fixed a problem where changes to attributes of
child DomNodes of circuit elements was not refreshing the property
editor. This enables the new dynamic property feature of the
CircuitEditor sample app.

ChildAttributePropertyDescriptor: fixed a problem with how the Equals()
method was implemented. Two ChildAttributePropertyDescriptor objects
should not be considered equivalent if they have different property
names or property categories. Updated unit tests to reflect this fix.

MultiPropertyDescriptor: exposed the property descriptors that are
inside this collection, so that the dynamic property feature of the
CircuitEditor sample app could work on a selection of multiple circuit
elements.