TreeControl user node re-ordering support
If a TreeControl’s ShowDragBetweenCue is set to true, and if the
context can be adapted to the new interface, IOrderedInsertionContext,
then the user will be allowed to re-order nodes by dragging. Also,
objects can be dragged into a specific location within a tree, like
when dragging from the palette.