PropertyGridView's Pick() could return wrong object
When multiple child collection editors are used in the Property Editor
(and so there are multiple child PropertyGridView objects), this Pick
method can return the wrong item. This was happening because
FindChildControls() returns the first instance of a PropertyGridView,
but there’s no ordering from top-to-bottom and the PropertyGridView may
not even be visible.